Moved tests for new angles module from test_convert to test_angles. Moved more functions over from convert to angles. Added __round__ method to classes.
